installer: set better mcollective timeouts
Bug 1122872 - "oo-mco" does not timeout when no ActiveMQ available
https://bugzilla.redhat.com/show_bug.cgi?id=1122872

Bug 1065048 - broker does not handle activemq outages gracefully
https://bugzilla.redhat.com/show_bug.cgi?id=1065048

When ActiveMQ is down or unreachable, with the previous default settings,
broker tools that operate via mcollective simply hang indefinitely.
With these defaults, mcollective queries will fail with useful error
after 6.3 seconds of retries, and a link is given to display options if
an administrator would like to adjust the behavior.

Additionally, the node interval for retrying a connection to ActiveMQ
was decreased to 4 seconds instead of the default 30. This way, nodes
will be visible faster after ActiveMQ returns.