Some output improvements; fix client config a little
Log now has a "Notice" level which is same as Info, but used for
highlighting process points as opposed to internal diagnostics output.
Also gave diagnostics a formal type; will help improve output etc.

Client config is a mess; this change pulls in origin code for getting
at what is actually the current config, but still does manual reading
and analysis of a .kubeconfig file. All of the origin flags that would
normally affect how the client reaches the master are added as "global"
flags.