revamp client config and discovery
In order to discover the client resources available, this now iterates
through the contexts in the RawConfig, generating a factory for each and
determining what projects are available. The primary goal is to find a
context that has cluster-admin access, or failing that, any access at
all. This will be necessary for determining what else can be diagnosed.